cpu/numa: fix failure when hot-remove cpu
When hot-remove cpu, the map from cpu to numa will set to NUMA_NO_NODE
which will lead to failure as the map is used by others. Thus we need a
specific map to descrip the unpluged cpu.

Here we introduce a new map to descrip the unpluged cpu map.

Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>