Debug: enable kernel boot up time measurement
To measure kernel boot up time, reserved region in pl011 is used as trap
MMIO region to record kernel boot timestamp.
For now, we trap at 2 point: the first is nearly the first instruction of
kernel start; the second is just before call the init bin for userspace.
These 2 points can cover the whole kernel boot time.

If there is no pl011, the first write is still there but the second
won't happen.

Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>